requirements.txt
discord.py==1.7.1
wavelink==0.9.10